dev-requirements.txt
# Debugging
# -------------------------------------
django-debug-toolbar==1.4
ipdb==0.8.1
ipython==4.0.1

# Testing and coverage
# -------------------------------------
pytest-django==2.9.1
pytest-pythonpath==0.7
pytest-cov==2.2.0
factory_boy==2.6.0
pdbpp==0.8.3
mock==1.3.0
flake8==2.5.1
django-extensions == 1.6.1
